fixes to the setup script
- don't run if not running from the correct directory
- if re-run, don't toss multiple lines into bash_profile
- if /usr/local/bin/setup.sh doesn't exist, don't try to run it
- homebrew won't bo on 2.7.14 forever, just assume homebrew is installed
  correctly and use 2.7 from the path.

+++ scripts/setup.sh
@@ -1,11 +1,21 @@
 #!/usr/bin/env bash
 
+if [ ! -e insecure_key ]; then
+    echo "Could not find ./insecure_key."
+    echo "Please run this script from the base of devlandia."
+    exit 1
+fi
+
+set -x
+
 pip install virtualenv virtualenvwrapper
-echo "WORKON_HOME="~/.virtualenvs"" >> ~/.bash_profile
-echo "source /usr/local/bin/setup.sh" >> ~/.bash_profile
+
+grep "WORKON_HOME" ~/.bash_profile || echo "WORKON_HOME="~/.virtualenvs"" >> ~/.bash_profile
+if [ -f /usr/local/bin/setup.sh ]
+   grep "/usr/local/bin/setup.sh" ~/.bash_profile || echo "source /usr/local/bin/setup.sh" >> ~/.bash_profile
 source $HOME/.bash_profile
 chmod 400 insecure_key && cp insecure_key ~/.ssh/
-mkvirtualenv devlandia -p /usr/local/Cellar/python/2.7.14/bin/python
+mkvirtualenv devlandia -p python2.7
 echo "cd $PWD" >> ~/.virtualenvs/devlandia/bin/postactivate
 source $HOME/.virtualenvs/devlandia/bin/activate
 pip install -r requirements.txt
